Siirretty kaikki classit (ja myös groupit) omiin tiedostoihinsa
