+++ AUTSball.py
@@ -2,9 +2,13 @@
 import pygame
 import math
 import sys
-import game_object
-import vector
 import music
+import groups
+import level
+import player
+import ball
+import text
+
 from colors import *
 
 
@@ -26,7 +30,7 @@ class AUTSBallGame:
 
         # Taustamusiikki
         self.music_player = music.MusicPlayer(screen='game', window_size=(self.screen_size_x, self.screen_size_y),
-                                              pos='bottomleft', group=TextGroup, shuffle=0)
+                                              pos='bottomleft', group=groups.TextGroup, shuffle=0)
         self.music_player.play()
 
         # SFX
@@ -40,10 +44,10 @@ class AUTSBallGame:
 
         # TODO: tähän assettien esilataus
         # Instansioidaan leveli, tämä lataa myös level-kuvan joka voi olla iiisooo
-        self.current_level = Level()
+        self.current_level = level.Level()
         # Instansioidaan pelaaja ja pallo
-        self.player = { 0: PlayerSprite(level=self.current_level, parent=self) }
-        self.ball = BallSprite(level=self.current_level, parent=self)
+        self.player = { 0: player.PlayerSprite(level=self.current_level, parent=self) }
+        self.ball = ball.BallSprite(level=self.current_level, parent=self)
 
         self.viewscreen_rect = None
         self.background_view_rect = None
@@ -90,11 +94,11 @@ class AUTSBallGame:
                                          self.screen_size_y)
 
             # Spritejen päivitykset tässä
-            BulletGroup.update(self.viewscreen_rect)
-            BallGroup.update(self.viewscreen_rect)
-            PlayerGroup.update()
-            EffectGroup.update()
-            TextGroup.update()
+            groups.BulletGroup.update(self.viewscreen_rect)
+            groups.BallGroup.update(self.viewscreen_rect)
+            groups.PlayerGroup.update()
+            groups.EffectGroup.update()
+            groups.TextGroup.update()
 
             # Päivitetään graffat vaan joka toisessa framessa
             if self.frame_counter % 2 == 0:
@@ -106,7 +110,6 @@ class AUTSBallGame:
         if self.quit_game:
             self.exit()
 
-
     def update_graphics(self):
         """ Grafiikoiden päivitysmetodi """
 
@@ -116,17 +119,17 @@ class AUTSBallGame:
         self.win.blit(self.current_level.image, self.background_view_rect)
 
         # Bullettien, pelaajan, pallon piirrot
-        BulletGroup.draw(self.win)
-        BallGroup.draw(self.win)
-        PlayerGroup.draw(self.win)
-        EffectGroup.draw(self.win)
-        TextGroup.draw(self.win)
+        groups.BulletGroup.draw(self.win)
+        groups.BallGroup.draw(self.win)
+        groups.PlayerGroup.draw(self.win)
+        groups.EffectGroup.draw(self.win)
+        groups.TextGroup.draw(self.win)
 
         # HUD
         # self.show_text((10, 10), "Speed: " + str(math.hypot(self.player[0].vx, self.player[0].vy)))
-        self.show_text((10, 50), "FPS: " + str(self.clock.get_fps()))
-        self.show_text((10, 10), str(self.score_green), color=GREEN, font_size=40)
-        self.show_text((750, 10), str(self.score_red), color=RED, font_size=40)
+        text.show_text(self.win, (10, 50), "FPS: " + str(self.clock.get_fps()))
+        text.show_text(self.win, (10, 10), str(self.score_green), color=GREEN, font_size=40)
+        text.show_text(self.win, (750, 10), str(self.score_red), color=RED, font_size=40)
 
         # Näytetään pallonsuuntamarkkeri
         # TODO: muuta pallon sijaan nuoli joka osoittaa oikeaan suuntaan
@@ -151,7 +154,7 @@ class AUTSBallGame:
             self.score_green += 1
             goal_text_color = GREEN
             self.goal_green_sound.play()
-        DisappearingText(pos=self.screen_center_point, text="GOAL!!!", frames_visible=120,
+        text.DisappearingText(pos=self.screen_center_point, text="GOAL!!!", frames_visible=120,
                          color=goal_text_color, font_size=120, flashes=1)
 
     def get_ball_angle_in_radians(self, ball):
@@ -162,477 +165,12 @@ class AUTSBallGame:
         y_difference = point1[1] - point2[1]
         return math.atan2(y_difference, x_difference)
 
-    def show_text(self, pos, text, color=(255, 255, 255), bgcolor=(0, 0, 0), font_size=24):
-        """ Utility-metodi tekstin näyttämiseen ruudulla """
-        font = pygame.font.Font(None, font_size)
-        textimg = font.render(text, 1, color, bgcolor)
-        self.win.blit(textimg, pos)
-
-    def empty_groups(self):
-        LevelGroup.empty()
-        BallGroup.empty()
-        PlayerGroup.empty()
-        BulletGroup.empty()
-        EffectGroup.empty()
-        TextGroup.empty()
 
     def exit(self):
         """ Tähän voi laittaa jotain mitä tulee ennen poistumista """
         pygame.quit()
         # Jostain syystä vaatii myös tämän, muuten jää infinite looppi taustalle vaikka pygame-ikkuna katoaakin
         sys.exit()
-
-# Sprite-ryhmät
-PlayerGroup = pygame.sprite.Group()
-LevelGroup = pygame.sprite.Group()
-BulletGroup = pygame.sprite.Group()
-BallGroup = pygame.sprite.Group()
-EffectGroup = pygame.sprite.Group()
-TextGroup = pygame.sprite.Group()
-
-
-class Level(pygame.sprite.Sprite):
-    """ Level-classi. Käytännössä vain taustakuva, logiikka tapahtuu muualla. """
-    def __init__(self, image_file=None, image=None):
-        pygame.sprite.Sprite.__init__(self, LevelGroup)
-
-        # Level-imagen lataus
-        if image is not None:
-            self.image = image
-        elif image_file is not None:
-            self.image = pygame.image.load(image_file).convert_alpha()
-        else:
-            self.image = pygame.image.load('gfx/test_arena_vertical_challenge.png').convert_alpha()
-
-        self.size_x = self.image.get_width()
-        self.size_y = self.image.get_height()
-        self.rect = self.image.get_rect()
-        self.center_point = self.size_x // 2, self.size_y // 2
-
-
-class EffectSprite(game_object.GameObject):
-    """ Yleinen efektisprite, tällä hetkellä tosin vain moottorin liekit """
-    def __init__(self, image=None, attached_player=None, attached_ball=None, effect_type=None, visible=1, parent=None):
-        game_object.GameObject.__init__(self, group=EffectGroup, image=image)
-        self.attached_player = attached_player
-        self.attached_ball = attached_ball
-        self.effect_type = effect_type
-        self.visible = visible
-
-    def update(self):
-        if self.visible:
-            player_dir_radians = math.radians(self.attached_player.heading)
-            dx = int(12 * math.sin(player_dir_radians))
-            dy = int(12 * math.cos(player_dir_radians))
-            self.rect.center = self.attached_player.rect.center[0] + dx, self.attached_player.rect.center[1] + dy
-            self.rot_self_image_keep_size(self.attached_player.heading)
-
-        else:
-            # jos ei visible niin heitetään vaan jonnekin kuuseen
-            self.rect.center = -100, -100
-
-    def destroy(self):
-        self.attached_player = None
-        self.kill()
-
-
-class BallSprite(game_object.GameObject):
-    """ Pallo. Osaa liittää itsensä pelaajaan ja poistaa liitoksen. """
-    def __init__(self, level=None, parent=None):
-        game_object.GameObject.__init__(self, group=BallGroup, image_file='gfx/ball_50_red.png', level=level, parent=parent)
-        self.start_position = self.level.center_point
-        self.x, self.y = self.start_position
-
-        # Player attachment
-        self.attached_player = None
-        self.attached_player_max_distance = 50  # "tetherin" pituus
-        self.attached_player_max_distance_squared = self.attached_player_max_distance**2  # distance-laskelmia varten
-
-        self.mass = 1.0
-        self.max_speed = 10
-
-        # Tämä tekee sen että tarkistetaan törmäys maaliin
-        self.is_ball = 1
-
-        # SFX
-        self.wall_collide_sound = pygame.mixer.Sound(file='sfx/thump4.wav')
-        self.bullet_collide_sound = pygame.mixer.Sound(file='sfx/metal_thud_3.wav')
-
-    def update(self, viewscreen_rect):
-        """ Päivittää palloa. Vaatii viewscreen_rect:in että osaa laskea näyttämisen oikein. """
-        self.viewscreen_rect = viewscreen_rect
-
-        # Jos törmää pelaajaan niin liitetään siihen
-        # Vain jos ei jo ole liitettynä!
-        if self.attached_player is None:
-            collide_list = pygame.sprite.spritecollide(self, PlayerGroup, dokill=False, collided=pygame.sprite.collide_circle)
-            if len(collide_list) > 0:
-                self.attach_to_player(collide_list[0])
-
-        # Jos on liitetty pelaajaan ja jos on liian kaukana niin vetävät toisiaan puoleensa
-        # Suht hyvä, voi viilata jos saa vielä paremmaksi
-        # TODO: weightit ei tunnu vaikuttavan?
-        # TODO: graffat tetherille
-        if self.attached_player is not None:
-            distance_to_player = self.distance(self.attached_player)
-            if distance_to_player >= self.attached_player_max_distance:
-                player_angle = game_object.get_angle_in_radians((self.attached_player.x, self.attached_player.y),
-                                                                 (self.x, self.y))
-
-                pull_vector_speed = (distance_to_player - self.attached_player_max_distance) * 0.02
-
-                ball_pull_vector = vector.MoveVector(speed=pull_vector_speed, direction=player_angle)
-                # Tässä rikotaan voiman ja vastavoiman lakia mutta who cares! (Newton pyörii haudassaan)
-                player_pull_vector = vector.MoveVector(speed=pull_vector_speed * -1 * 0.4, direction=player_angle)
-
-                self.move_vector.add_vector(ball_pull_vector)
-                self.attached_player.move_vector.add_vector(player_pull_vector)
-
-        self.update_movement()
-
-        self.check_out_of_bounds()
-        self.check_collision_with_wall_and_goal()
-        self.check_collision_with_bullets(BulletGroup)
-
-    def collide_tether(self, other_object):
-        """ 
-        Eksperimentaalinen tetherin törmäysmetodi. Ei toimi niin kuin haluaisin.
-        Ideana tässä että tether-collide on oletettavasti ekvivalentti normaalille törmäykselle niin että
-        objektien paikat vaihdetaan
-        Ei tällä hetkellä käytössä
-        """
-        new_self = other_object
-        new_other = self
-        angle_to_other = game_object.get_angle_in_radians(new_other.rect.center, new_self.rect.center)
-        new_self.move_vector.set_direction(angle_to_other - math.pi)
-        new_other.move_vector.set_direction(angle_to_other)
-
-        speed1 = new_self.move_vector.get_speed()
-        speed2 = new_other.move_vector.get_speed()
-        mass1 = new_self.mass
-        mass2 = new_other.mass
-        speed1_new = (mass2 / mass1) * speed2
-        speed2_new = (mass1 / mass2) * speed1
-        self.move_vector.set_speed(speed2_new * -1)
-        other_object.move_vector.set_speed(speed1_new * -1)
-
-    def reset(self):
-        """ 
-        HUOM! Tämä metodi overrideaa GameObjectin reset-metodin että osaa detachata pallon pelaajasta.
-        Siksi kutsuu GameObjectin resetin näin käsin.
-        """
-        game_object.GameObject.reset(self)
-        self.detach()
-
-    def shoot(self, direction=0, speed=0, x=None, y=None):
-        """ 
-        Ampuu itsensä määritettyyn suuntaan, määritetyllä nopeudella, alkaen määritetyistä koordinaateista.
-        Tätä kutsuu PlayerSpriten shoot-metodi, joka hoitaa detachauksen ja antaa tarvittavat tiedot
-        """
-
-        if x is not None and y is not None:
-            # Tämä tehdään jos annettu uudet koordinaatit
-            # Jostain syystä vaatii direktion korjauksen tässä
-            self.move_vector.set_speed_direction(speed, math.radians(270 - direction))
-            self.x = int(x)
-            self.y = int(y)
-            self.update_rect()
-        else:
-            # Jos ei ole annettu uusia koordinaatteja niin lisätään vain liikevektoriin määritetyt
-            # direction ja speed
-            self.move_vector.add_vector(vector.MoveVector(speed=speed, direction=math.radians(270 - direction)))
-
-
-    def attach_to_player(self, player):
-        """ 
-        Tämä metodi liittää pallon pelaajaan. Olisi tarkoitus myös lisätä painoa mutta paino ei toimi
-        oikein vielä.
-        """
-        self.attached_player = player
-        player.attach_ball(self)
-        # self.tether = EffectSprite(image=pygame.Surface((0,0)), effect_type='tether',
-        #                            attached_ball=self, attached_player=player, parent=self.parent)
-        # TODO: korjaa weight että tämä voidaan enabloida
-        # self.attached_player.weight += self.weight
-
-    def detach(self):
-        """ Tämä metodi poistaa liitoksen pelaajaan. """
-        # self.attached_player.weight -= self.weight
-        # print("Ball detach method called. Attached player:", self.attached_player)
-        if self.attached_player is not None:
-            self.attached_player.detach()
-            self.attached_player = None
-            # self.tether.kill()
-            # self.tether = None
-
-
-class BulletSprite(game_object.GameObject):
-    """ direction asteina, tulee PlayerSpriten headingista """
-    def __init__(self, parent=None, level=None, x=0, y=0, direction=0, parent_speed=0, speed=5, type='basic'):
-        game_object.GameObject.__init__(self, group=BulletGroup, image_file='gfx/bullet_5.png', start_position=(x, y),
-                                        level=level, parent=parent)
-        self.rect.center = (x, y)
-        self.move_vector.set_speed_direction(speed, math.radians(270 - direction))
-        self.max_speed = 20
-        self.explosion_force = 1
-
-        self.is_bullet = 1
-
-        # SFX
-        self.wall_collide_sound = pygame.mixer.Sound(file='sfx/thump3.wav')
-        self.wall_collide_sound.set_volume(1)
-
-    def update(self, viewscreen_rect):
-        self.viewscreen_rect = viewscreen_rect
-        self.update_movement()
-        self.check_out_of_bounds()
-        # print("Speed:", self.move_vector.get_speed())
-        # Tehdään nämä vain jos on olemassa
-        if self in BulletGroup:
-            self.check_collision_with_wall_and_goal()
-            if self in BulletGroup:
-                if self.speculate_collision_with_wall() == 1:
-                    # print(self.move_vector.get_speed())
-                    # Vähennetään nopeutta jos spekulointi havaitsee törmäyksen
-                    self.move_vector.set_speed(min(self.move_vector.get_speed(), 3))
-                self.update_rect()
-
-    def check_out_of_bounds(self):
-        """ Overrideaa GameObjectin metodin koska pitää tuhota bulletti jos on out of bounds """
-        if self.x < 0 or self.y < 0 or self.x >= self.level.size_x or self.y >= self.level.size_y:
-            self.x = 0
-            self.y = 0
-            self.kill()
-
-
-class PlayerSprite(game_object.GameObject):
-    def __init__(self, level=None, parent=None):
-        # Lisätään PlayerGroup-ryhmään
-        game_object.GameObject.__init__(self, group=PlayerGroup, level=level, parent=parent,
-                                        image_file='gfx/ship1_red_20px.png')
-
-        # Graffat
-        self.motor_flame_image = pygame.image.load('gfx/motor_flame_10.png').convert_alpha()
-        self.thrust_gfx = EffectSprite(attached_player=self, image=self.motor_flame_image,
-                                       effect_type='motorflame', visible=0, parent=parent)
-        self.rect.center = self.parent.screen_center_point
-        self.is_centered_on_screen = 1
-
-        # Sound effex
-        self.motor_sound = pygame.mixer.Sound(file='sfx/shhhh_v2.wav')
-        # self.motor_sound.set_volume(0.4)
-        self.motor_sound_playing = 0
-        self.bullet_sound = pygame.mixer.Sound(file='sfx/pop.wav')
-        self.ball_shoot_sound = pygame.mixer.Sound(file='sfx/pchou.wav')
-        self.ball_capture_sound = pygame.mixer.Sound(file='sfx/ball_capture.wav')
-        self.wall_collide_sound = pygame.mixer.Sound(file='sfx/thump4.wav')
-        self.bullet_collide_sound = pygame.mixer.Sound(file='sfx/metal_thud_2.wav')
-
-        # Koordinaatit
-        self.start_position = (700, 1200)
-        self.x, self.y = self.start_position
-        self.x_previous, self.y_previous = self.x, self.y
-
-        # Heading ja thrust
-        self.heading = 0  # HUOM! Heading asteina koska Pygame käyttää niitä rotaatioissa
-        self.thrust = 0
-
-        # Pallo
-        self.attached_ball = None
-
-        # Shipin ominaisuudet
-        self.handling = int(5)  # kuinka monta astetta kääntyy per frame
-        self.max_thrust = 0.35  # kun FPS 60, gravity 0.1 ja mass 1 niin 0.35 on aika hyvä
-        self.max_speed = 10
-        self.mass = 1.0
-        self.cooldown_basic_shot = 5 # framea
-        self.cooldown_after_ball_shot = 60 # cooldown sen jälkeen kun pallo on ammuttu
-        self.cooldown_counter = 0 # cooldown-counter1
-        self.recovery_time = 3  # sekunteja jopa!
-        self.recovery_started_at = 0
-
-    def update(self):
-        # Lisätään liikemäärään thrust-vektori
-        # Tässä jopa ottaa jo massan huomioon!
-        self.move_vector.add_to_vx((self.thrust / self.mass * math.sin(math.radians(self.heading)) * -1))
-        self.move_vector.add_to_vy((self.thrust / self.mass * math.cos(math.radians(self.heading)) * -1))
-
-        self.update_movement()
-
-        self.check_out_of_bounds()
-        self.check_collision_with_wall_and_goal()
-        self.check_collision_with_players(PlayerGroup)
-        self.check_collision_with_bullets(BulletGroup)
-
-        # Lasketaan cooldownia
-        if self.cooldown_counter > 0:
-            self.cooldown_counter -= 1
-
-        # Jos on pallo kytkettynä niin lisätään paljon cooldownia
-        if self.attached_ball is not None:
-            self.cooldown_counter = self.cooldown_after_ball_shot
-
-        if self.recovery_started_at != 0:
-            if (pygame.time.get_ticks() - self.recovery_started_at) // 1000 > self.recovery_time - 1:
-                self.reset()
-                self.recovery_started_at = 0
-
-    def attach_ball(self, ball):
-        if self.attached_ball is None:
-            self.attached_ball = ball
-            self.force_play_sound(self.ball_capture_sound)
-
-    def detach(self):
-        self.attached_ball = None
-        self.radius = self.original_radius
-
-    def accelerate(self):
-        self.thrust = self.max_thrust
-        self.thrust_gfx.visible = 1
-        if not self.motor_sound_playing:
-            if self.motor_sound is not None:
-                self.force_play_sound(self.motor_sound, -1)
-                self.motor_sound_playing = 1
-        # print(self.motor_sound.get_num_channels())
-        # print(pygame.mixer.get_busy())
-        # if self.motor_sound.get_num_channels() == 0:
-        #     pygame.mixer.find_channel(True).play(self.motor_sound, -1)
-
-    def stop_acceleration(self):
-        self.thrust = 0
-        self.thrust_gfx.visible = 0
-        self.motor_sound.stop()
-        self.motor_sound_playing = 0
-
-    def rotate_right(self):
-        self.heading -= self.handling
-        if self.heading < 0:
-            self.heading += 360
-        # TODO: laske rotaatiot latausvaiheessa valmiiksi
-        self.rot_self_image_keep_size(self.heading)
-
-    def rotate_left(self):
-        self.heading += self.handling
-        if self.heading > 360:
-            self.heading -= 360
-        self.rot_self_image_keep_size(self.heading)
-
-    def shoot(self):
-        # Ammutaan perusammus
-        # Pelaajan nopeus vaikuttaa ammuksen vauhtiin
-        # TODO: pelaajan nopeus lisää aina ammuksen nopeutta saman verran riippumatta siitä mihin suuntaan se ammutaan!
-        # Asetetaan ammuksen alkupiste riittävän kauas pelaajasta ettei törmää saman tien siihen
-        if self.cooldown_counter == 0:
-            self.force_play_sound(self.bullet_sound)
-            bullet_x = int(10 * math.sin(math.radians(self.heading)) * -1 + self.x)
-            bullet_y = int(10 * math.cos(math.radians(self.heading)) * -1 + self.y)
-            BulletSprite(level=self.level, parent=self.parent, x=bullet_x, y=bullet_y, direction=self.heading,
-                         speed=10 + self.move_vector.get_speed())
-            self.cooldown_counter = self.cooldown_basic_shot
-
-        # Jos pallo on liitettynä niin ammutaan se
-        if self.attached_ball is not None:
-            # ball_x = self.attached_ball.image.get_width() * math.sin(math.radians(self.heading)) * -1 + self.x
-            # ball_y = self.attached_ball.image.get_height() * math.cos(math.radians(self.heading)) * -1 + self.y
-
-            # self.attached_ball.shoot(x=ball_x, y=ball_y, direction=self.heading, speed=10)
-            self.attached_ball.shoot(direction=self.heading, speed=10)
-            self.attached_ball.detach()
-            self.force_play_sound(self.ball_shoot_sound)
-
-    def recover(self):
-        """ Aloittaa recovery-laskennan """
-        self.recovery_started_at = pygame.time.get_ticks()
-        DisappearingText(pos=self.parent.screen_center_point, text="RECOVERING...", frames_visible=240, flashes=1,
-                         font_size=80, color=RED)
-
-class DisappearingText(pygame.sprite.Sprite):
-    """ Näyttää ruudulla tekstin x framen ajan """
-    def __init__(self, pos=(0,0), text="", frames_visible=60,
-                 color=WHITE, bgcolor=None, font_size=24, flashes=0, flash_interval=10):
-        pygame.sprite.Sprite.__init__(self, TextGroup)
-
-        self.frame_counter = 0
-        self.frames_visible = frames_visible
-
-        font = pygame.font.Font(None, font_size)
-        self.image = font.render(text, 1, color, bgcolor)
-        self.original_position = pos
-        self.rect = self.image.get_rect()
-        self.rect.center = pos
-
-        self.flashes = flashes
-        self.flash_interval = flash_interval
-        self.visible = 1
-
-    def update(self):
-        self.frame_counter += 1
-        if self.frame_counter > self.frames_visible:
-            self.kill()
-
-        if self.flashes and self.frame_counter % self.flash_interval == 0:
-            self.toggle_image()
-
-    def toggle_image(self):
-        if self.visible:
-            self.visible = 0
-            self.rect.center = -100, -100
-        else:
-            self.visible = 1
-            self.rect.center = self.original_position
-
-
-class ScrollingText(pygame.sprite.Sprite):
-    """ Scrollaa tekstiä ruudulla """
-    def __init__(self, y_pos=0, screen_size_x=800, text="", scroll_direction='left', scroll_speed=5,
-                 color=WHITE, bgcolor=None, font_size=24, flashes=0, flash_interval=10):
-        pygame.sprite.Sprite.__init__(self, TextGroup)
-
-        self.frame_counter = 0
-
-        font = pygame.font.Font(None, font_size)
-        self.image = font.render(text, 1, color, bgcolor)
-        self.original_image = self.image
-        self.empty_image = pygame.Surface((0, 0))
-        self.rect = self.image.get_rect()
-        self.screen_size_x = screen_size_x
-
-        self.scroll_direction = scroll_direction
-
-        if scroll_direction == 'left':
-            self.rect.midleft = screen_size_x, y_pos
-            self.original_position = self.rect.midleft
-            self.scroll_speed = scroll_speed * -1
-        else:
-            self.rect.midright = 0, y_pos
-            self.original_position = self.rect.midright
-            self.scroll_speed = scroll_speed
-
-        self.flashes = flashes
-        self.flash_interval = flash_interval
-        self.visible = 1
-
-    def update(self):
-        self.rect.x += self.scroll_speed
-        if self.scroll_direction == 'left':
-            if self.rect.midright[0] < 0:
-                self.rect.midleft = self.original_position
-        else:
-            if self.rect.midleft[0] > self.screen_size_x:
-                self.rect.midright = self.original_position
-
-        if self.flashes:
-            self.frame_counter += 1
-            if self.frame_counter % self.flash_interval == 0:
-                self.toggle_image()
-
-    def toggle_image(self):
-        if self.visible:
-            self.visible = 0
-            self.image = self.empty_image
-        else:
-            self.visible = 1
-            self.image = self.original_image
 
 
 if __name__ == '__main__':
