+++ level.py
@@ -1,0 +1,23 @@
+# -*- coding: utf8 -*-
+import pygame
+import groups
+
+
+class Level(pygame.sprite.Sprite):
+    """ Level-classi. Käytännössä vain taustakuva, logiikka tapahtuu muualla. """
+    def __init__(self, image_file=None, image=None, group=groups.LevelGroup):
+        pygame.sprite.Sprite.__init__(self, group)
+
+        # Level-imagen lataus
+        if image is not None:
+            self.image = image
+        elif image_file is not None:
+            self.image = pygame.image.load(image_file).convert_alpha()
+        else:
+            self.image = pygame.image.load('gfx/test_arena_vertical_challenge.png').convert_alpha()
+
+        self.size_x = self.image.get_width()
+        self.size_y = self.image.get_height()
+        self.rect = self.image.get_rect()
+        self.center_point = self.size_x // 2, self.size_y // 2
+
